+++ pyhash.c
@@ -1817,7 +1817,7 @@ static unsigned long hashtable_kilos;
 unsigned long allochash(unsigned long nr_kilos)
 {
 #if defined(FXF)
-  unsigned long const one_kilo = 1<<10;
+  size_t const one_kilo = 1<<10;
   while (fxfInit(nr_kilos*one_kilo)==-1)
     /* we didn't get hashmemory ... */
     nr_kilos /= 2;
